+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.1.0 - 2017-05-14
+### Fixed
+- Database file being saved at the wrong place
+
+### Added
+- Can now specify `all` as a file type to dowload all file types (Except `iso`)
+- Pass the file containing the `s_session_id` with the `-s` flag
+- Update the database file with `-u`
+
 ## 1.0.4 - 2017-05-12
 ### Fixed
 - Incorrect data type returned when session token could not be loaded